+++ src/main/java/com/codetaylor/mc/pyrotech/modules/core/block/BlockRefractoryBricksStairs.java
@@ -1,0 +1,15 @@
+package com.codetaylor.mc.pyrotech.modules.core.block;
+
+import net.minecraft.block.BlockStairs;
+import net.minecraft.block.state.IBlockState;
+
+public class BlockRefractoryBricksStairs
+    extends BlockStairs {
+
+  public static final String NAME = "refractory_brick_stairs";
+
+  public BlockRefractoryBricksStairs(IBlockState modelState) {
+
+    super(modelState);
+  }
+}